+++ CHANGELOG.rst
@@ -60,6 +60,7 @@ support the `disconnected` message before upgrading.
 
 .. _SaltyRTC 1.0 Protocol: https://github.com/saltyrtc/saltyrtc-meta/blob/protocol-1.0/Protocol.md
 
+.. _Unreleased: https://github.com/saltyrtc/saltyrtc-server-python/compare/v3.0.1...HEAD
 .. _3.0.1: https://github.com/saltyrtc/saltyrtc-server-python/compare/v3.0.0...v3.0.1
 .. _3.0.0: https://github.com/saltyrtc/saltyrtc-server-python/compare/v2.0.1...v3.0.0
 .. _2.0.1: https://github.com/saltyrtc/saltyrtc-server-python/compare/v2.0.0...v2.0.1